Package Role Information in BagIt Exports (#53)
Resolves: https://jira.lyrasis.org/browse/DS-4481

* Adds Pojos for tags defined in the DSpaceRoles schema
* Add a utility class for retrieving and ingesting DSpaceRoles from DSpaceObjects
* Package DSpaceRoles as roles.xml for BagIt exports
* Read DSpaceRoles during import and ingest using the RoleIngester

Notes:

The work for retrieving roles for DSpaceObjects was taken heavily from the RoleDisseminator and updated to fit in a way which works with how xml writing is handled for BagIt exports. Since the xml schema for the DSpaceRoles is shared between implementations, the RoleIngester is used in order to provide ingestion.

There is a difference in that BagIt implementation will not emit passwords for any EPerson as there isn't a passwords option in the RTS configuration. If needed this can be added on to the current implementation.