Minimize use of Curator.curationContext() and re-enable tests
This was done to improve unit testability of the code.

Context instances can be passed into most constructors or methods under test now,
which reduces the need to mock other, larger parts of the DSpace Kernel
in order to perform unit tests.

After this change, the only place this static method is called is in the
curation task implementations' perform(DSpaceObject) methods. It is necessary
here because the signature of perform(DSpaceObject) is defined in DSpace itself,
and does not currently pass in a context to use.