Update sample curate.cfg for DSpace 7
This re-creates curate.cfg based on the out-of-box curate.cfg
for DSpace 7, adding only those configurations that are relevant
for RTS in DSpace 7.

Notably, the ui-related configurations have been removed due
to curation tasks now executing from the scripts and processes
functionality: https://github.com/DSpace/DSpace/pull/2820

In addition, this corrects an error in the sample (commented) configs
for bagit and checkm, where the plugin syntax was not updated to
use the new way of specifying multiple values as of DSpace 6.